refactor: split useMemoryAutomation into sub-modules (T9)
useMemoryAutomation (706 → ~540 lines) decomposed into:
- memoryDraftContext.js: pure context-building functions
  (buildMemoryContinuityContext, buildMemoryDraftLoreContext,
  buildMemoryDraftSummaryExcerpt) — zero dependencies, ~75 lines
- useMemoryBatchGeneration.js: batch generation composable
  (runBatchDraftGeneration, runBatchDraftGenerationFromIds,
  generateSingleDraft, handleMemoryBatchGenerate) — ~170 lines
- useMemoryAutomation.js: core automation orchestrator
  (createPendingMemoryDraft, generateMemoryDraftForMessages,
  runMemoryAutomationAfterStableTurn, bootstrapImportedMemoryDrafts,
  handleMemoryQuickModelChange) — composes from sub-modules

Public API unchanged — ChatView.vue imports unchanged.
